% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_neig.r
\name{create.neig}
\alias{create.neig}
\title{create.neig}
\usage{
create.neig(data.set, nb.nn = 4, duplicate = "remove", verbose = FALSE)
}
\description{
Initiates the neighbourhood relationships between the points in the processed data set
}
\author{
Pierre Roudier
}

