% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract_extrema.r
\name{extract.extrema}
\alias{extract.etxrema}
\alias{extract.extrema}
\title{extract.extrema}
\usage{
extract.extrema(tri.obj, n.extrema.min = 1)
}
\description{
Explicitely returns a list with the tri objects of the extrema
}
\author{
Pierre Roudier
}

