% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extrema_irr.r
\name{extrema.irr}
\alias{extrema.irr}
\title{extrema.irr}
\usage{
extrema.irr(data.set, gridded.data, neig = NULL, zcol = "z",
  duplicate = "remove", nb.nn = 4, thresh.extrema = 1, verbose = FALSE)
}
\description{
Finds regional extrema on a irregularly sampled data set
}
\author{
Pierre Roudier
}

