% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mean_enveloppe.r
\name{return.mean.enveloppe}
\alias{return.mean.enveloppe}
\title{return.mean.enveloppe}
\usage{
return.mean.enveloppe(extrema, donnee, zcol = "z", method = "splines",
  n.pts.spline = 3, verbose = TRUE)
}
\description{
Returns the mean enveloppe
}
\author{
Pierre Roudier
}

