% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_neig.r
\name{create.neig}
\alias{create.neig}
\title{create.neig}
\usage{
create.neig(data.set, nb.nn = 4, duplicate = "remove", verbose = FALSE)
}
\arguments{
\item{data.set}{Data set to create neighbourhood from.}

\item{nb.nn}{Number of nearest neighbours. Defaults to 4.}

\item{duplicate}{Ignored.}

\item{verbose}{Prints progress information messages. Defaults to FALSE.}
}
\description{
Internal function, initiates the neighbourhood relationships between the points in the processed data set.
}
\author{
Pierre Roudier
}
