% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_extrema.r
\name{extract.extrema}
\alias{extract.extrema}
\alias{extract.etxrema}
\title{extract.extrema}
\usage{
extract.extrema(tri.obj, n.extrema.min = 1)
}
\arguments{
\item{tri.obj}{.}

\item{n.extrema.min}{.}
}
\description{
Internal function, explicitely returns a list with the tri objects of the extrema.
}
\author{
Pierre Roudier
}
