% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrema_irr.r
\name{extrema.irr}
\alias{extrema.irr}
\title{extrema.irr}
\usage{
extrema.irr(data.set, gridded.data, neig = NULL, zcol = "z",
  duplicate = "remove", nb.nn = 4, thresh.extrema = 1, verbose = FALSE)
}
\arguments{
\item{data.set}{.}

\item{gridded.data}{.}

\item{neig}{Neighbourhood object.}

\item{zcol}{Name of the column containing the data.}

\item{duplicate}{What to do with duplicates. Defaults to `remove`.}

\item{nb.nn}{Number of nearest neighbours to take into account if data is on a grid. Defaults to 4.}

\item{thresh.extrema}{Significative threshold for the extrema. Defaults to 1.}

\item{verbose}{Prints progress information messages. Defaults to FALSE.}
}
\description{
Internal function, finds regional extrema on a irregularly sampled data set
}
\author{
Pierre Roudier
}
