% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spemd.r
\name{spEMD}
\alias{spEMD}
\title{spEMD}
\usage{
spEMD(data, zcol = "z", method = "splines", n.imf.max = 10,
  n.sp.max = 5, n.extrema.min = 1, stoprule = "mean.imf",
  stoprule.extrema = TRUE, thresh.extrema = 1, tol = 0,
  diff.nb.extrema = 0.05, abs.nb.extrema = 5, nb.nn = 4,
  n.pts.spline = 4, neig = NULL, save_neig = TRUE, verbose = TRUE)
}
\arguments{
\item{data}{.}

\item{zcol}{.}

\item{method}{.}

\item{n.imf.max}{.}

\item{n.sp.max}{.}

\item{n.extrema.min}{.}

\item{stoprule}{.}

\item{stoprule.extrema}{.}

\item{thresh.extrema}{.}

\item{tol}{.}

\item{diff.nb.extrema}{}

\item{abs.nb.extrema}{.}

\item{nb.nn}{.}

\item{n.pts.spline}{.}

\item{neig}{.}

\item{save_neig}{.}

\item{verbose}{.}

\item{file}{.}
}
\value{
.
}
\description{
2D EMD for spatial objects
}
\author{
Pierre Roudier
}

